+++ backend/app/alembic/script.py.mako
@@ -6,10 +6,10 @@ Create Date: ${create_date}
 
 """
 from typing import Sequence, Union
-import sqlmodel.sql.sqltypes
 
 from alembic import op
 import sqlalchemy as sa
+import pgvector.sqlalchemy
 ${imports if imports else ""}
 
 # revision identifiers, used by Alembic.
